successfully create spot with new fields and tags

+++ backend/db/migrations/20220627181005-create-spot.js
@@ -60,7 +60,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.INTEGER(6)
+        type: Sequelize.INTEGER
       },
       createdAt: {
         allowNull: false,
